Split lockfile code away from osmose_run.py
